sync filter params moved to matrix config
