+++ changelog.d/filter.sdk
@@ -1,0 +1,1 @@
+[Sync] Sync Filter params are moved to MatrixConfiguration and will not be stored in session realm to avoid bug when session cache is cleared
