+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/filter/FilterModule.kt
@@ -19,7 +19,6 @@ package org.matrix.android.sdk.internal.session.filter
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
-import org.matrix.android.sdk.api.session.sync.FilterService
 import org.matrix.android.sdk.internal.session.SessionScope
 import retrofit2.Retrofit
 
@@ -38,9 +37,6 @@ internal abstract class FilterModule {
 
     @Binds
     abstract fun bindFilterRepository(repository: DefaultFilterRepository): FilterRepository
-
-    @Binds
-    abstract fun bindFilterService(service: DefaultFilterService): FilterService
 
     @Binds
     abstract fun bindSaveFilterTask(task: DefaultSaveFilterTask): SaveFilterTask
