+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/sync/filter/SyncFilterBuilder.kt
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.api.session.sync.filter
+package org.matrix.android.sdk.internal.sync.filter
 
 import org.matrix.android.sdk.api.session.homeserver.HomeServerCapabilities
+import org.matrix.android.sdk.api.session.sync.filter.SyncFilterParams
 import org.matrix.android.sdk.internal.session.filter.Filter
 import org.matrix.android.sdk.internal.session.filter.RoomEventFilter
 import org.matrix.android.sdk.internal.session.filter.RoomFilter
-import org.matrix.android.sdk.internal.sync.filter.SyncFilterParams
 
-class SyncFilterBuilder {
+internal class SyncFilterBuilder {
     private var lazyLoadMembersForStateEvents: Boolean? = null
     private var lazyLoadMembersForMessageEvents: Boolean? = null
     private var useThreadNotifications: Boolean? = null
