--- matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/sync/filter/SyncFilterParams.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright 2022 The Matrix.org Foundation C.I.C.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.matrix.android.sdk.internal.sync.filter
-
-internal data class SyncFilterParams(
-        val lazyLoadMembersForStateEvents: Boolean? = null,
-        val lazyLoadMembersForMessageEvents: Boolean? = null,
-        val useThreadNotifications: Boolean? = null,
-        val listOfSupportedEventTypes: List<String>? = null,
-        val listOfSupportedStateEventTypes: List<String>? = null,
-)
