--- vector/src/test/java/im/vector/app/test/fakes/FakeFilterService.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2022 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package im.vector.app.test.fakes
-
-import io.mockk.coEvery
-import io.mockk.coVerify
-import io.mockk.just
-import io.mockk.mockk
-import io.mockk.runs
-import org.matrix.android.sdk.api.session.sync.FilterService
-import org.matrix.android.sdk.api.session.sync.filter.SyncFilterBuilder
-
-class FakeFilterService : FilterService by mockk() {
-
-    fun givenSetFilterSucceeds() {
-        coEvery { setSyncFilter(any()) } just runs
-    }
-
-    fun verifySetSyncFilter(filterBuilder: SyncFilterBuilder) {
-        coVerify { setSyncFilter(filterBuilder) }
-    }
-}
